refactor(tests): service health metrics test moved to tested mod
