+++ crates/api-core/src/logging/service_health_metrics.rs
@@ -124,3 +124,64 @@ pub(crate) fn start_export_service_health_metrics(health_context: ServiceHealthC
             .build();
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashMap;
+    use std::sync::{Arc, Mutex};
+
+    use carbide_utils::test_support::test_meter::TestMeter;
+    use model::resource_pool::ResourcePoolStats;
+    use prometheus_text_parser::ParsedPrometheusMetrics;
+    use sqlx::PgPool;
+
+    use super::{ServiceHealthContext, start_export_service_health_metrics};
+
+    #[crate::sqlx_test]
+    async fn test_service_health_metrics(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
+        let test_meter = TestMeter::default();
+        let context = ServiceHealthContext {
+            meter: test_meter.meter(),
+            database_pool: pool,
+            resource_pool_stats: Arc::new(Mutex::new(HashMap::from([
+                (
+                    "pool1".to_string(),
+                    ResourcePoolStats {
+                        used: 10,
+                        free: 20,
+                        auto_assign_free: 20,
+                        auto_assign_used: 10,
+                        non_auto_assign_free: 0,
+                        non_auto_assign_used: 0,
+                    },
+                ),
+                (
+                    "pool2".to_string(),
+                    ResourcePoolStats {
+                        used: 20,
+                        free: 10,
+                        auto_assign_free: 10,
+                        auto_assign_used: 20,
+                        non_auto_assign_free: 0,
+                        non_auto_assign_used: 0,
+                    },
+                ),
+            ]))),
+        };
+        start_export_service_health_metrics(context);
+
+        let expected_metrics = include_str!("test_data/test_service_health_metrics.txt")
+            .parse::<ParsedPrometheusMetrics>()
+            .unwrap()
+            .scrub_build_attributes();
+        let metrics = test_meter
+            .export_metrics()
+            .parse::<ParsedPrometheusMetrics>()
+            .unwrap()
+            .scrub_build_attributes();
+
+        assert_eq!(expected_metrics, metrics);
+
+        Ok(())
+    }
+}
